Made changes to the Assets, Bootsrap, HTML, and CSS
Removed the SVG icons in Assets and replaced them with SVG code in HTML.
Added startup docs to the  Assets.
Added SVG logo to the Image Assets.

Removed the CDN Bootsrap links and instead downloaded the min JS and CSS for flexibility when offline.

Modified the nav to header, customized responsiveness.
Added CTA and Footer in the body.
Added theme toggler for the page.